build: enable changing C++ compilation priority for builds by ninja
`configure.py` now handles a new flag `--compiler-niceness` which
accepts a number. The value is written into the file `build.ninja` and
influences the process priority of the subsequent C++ compilation tasks
launched by Ninja.

For example, in order to configure ninja to run the C++ compilations
with lower priority (not to block the system so much),
run `configure.py` with `--compiler-niceness 4`
For more information about the possible values, please inspect the
manual regarding the linux tool `nice`.

`install-dependencies.sh` requires now the installation of `coreutils`
package which contains the `nice` binary.