alternator: make listen socket backlog configurable and live-updatable
Add new config variable `alternator_listen_socket_backlog` that controls
the backlog parameter passed to listen() for Alternator's HTTP and HTTPS
server sockets. The default value is 1024 (previously the Seastar default
of 100 was used).

The parameter is live-updatable: when changed at runtime, the new backlog
is applied to all active Alternator server sockets by re-invoking the
listen() syscall on the underlying file descriptors (supported by Linux).

This includes a Seastar change that adds set_listen_backlog() to
server_socket and http_server, allowing the listen backlog to be
updated on already-listening sockets.

Fixes #18021